payment_api/requirements.txt
# payment_api/requirements.txt
fastapi
uvicorn[standard]
kafka-python